remove addvector with auxilary context.
TieredHNSWIndex::insertVectorToHNSW is now responsible to preprocess the blob according to the hnsw preprocessor

HNSW: introduce indexvector(blob, label, state) that inserts a stored vector to the graph.

Tiredindex: preprocessing for query is done according to the frontend index first.

Factory:
HNSW: added is_normalized arg to decide if we need IP or cosine